refactor: selection-sheet 컴포넌트 분리
